one song from local storage
added some micro interactions thanks to lottie and for now you can add only 1 song going to the second screen(local audio). This is the raw file, in next update it will be documented and somewhat I will try to optimise. Note: If you want to change song you have to close the app reopoen and do again