done for now
i have learned a lot and probably wasted a lot of time, but now gonna be serious. For this app, it is done for now. There are bugs, no further optimization and new playlist does not play. For now I will work on my next project of a trello twin and then probably get back to it after a shopping app with totally new ui, code, storage, and many more.